+++ demo/src/main/java/com/example/demo/repository/ClienteRepository.java
@@ -11,11 +11,11 @@ import com.example.demo.entitys.Cliente;
 @Repository
 public class ClienteRepository {
     
-    private final Map<Integer, Cliente> clientes = new HashMap<>();
+    private final Map<Long, Cliente> clientes = new HashMap<>();
 
     public ClienteRepository() {
-        clientes.put(1, new Cliente(
-        1,
+        clientes.put(1L, new Cliente(
+        1L,
         "Pablo",
         "García",
         "PabloGarcia21@gmail.com",
@@ -25,8 +25,8 @@ public class ClienteRepository {
         "3001234567"
     ));
 
-        clientes.put(2, new Cliente(
-                2,
+        clientes.put(2L, new Cliente(
+                2L,
                 "María",
                 "Gómez",
                 "maria.gomez@email.com",
@@ -36,8 +36,8 @@ public class ClienteRepository {
                 "3019876543"
         ));
 
-        clientes.put(3, new Cliente(
-                3,
+        clientes.put(3L, new Cliente(
+                3L,
                 "Andrés",
                 "Martínez",
                 "andres.martinez@email.com",
@@ -47,8 +47,8 @@ public class ClienteRepository {
                 "3024567890"
         ));
 
-        clientes.put(4, new Cliente(
-                4,
+        clientes.put(4L, new Cliente(
+                4L,
                 "Laura",
                 "Ramírez",
                 "laura.ramirez@email.com",
@@ -58,8 +58,8 @@ public class ClienteRepository {
                 "3106543210"
         ));
 
-        clientes.put(5, new Cliente(
-                5,
+        clientes.put(5L, new Cliente(
+                5L,
                 "Camilo",
                 "Torres",
                 "camilo.torres@email.com",
@@ -70,7 +70,7 @@ public class ClienteRepository {
         ));
     }
         
-    public Cliente findById(Integer id) {
+    public Cliente findById(Long id) {
         return clientes.get(id);
     }
 
@@ -79,13 +79,13 @@ public class ClienteRepository {
 
     }
 
-    public Integer count() {
-        return clientes.size();
+    public Long count() {
+        return (long) clientes.size();
     }
     public void save(Cliente cliente) {
         if(cliente.getId()== null){
-        Integer tam = count();
-        Integer lastid = clientes.get(tam).getId();
+        Long tam = count();
+        Long lastid = clientes.get(tam).getId();
         cliente.setId(lastid + 1);
         clientes.put(cliente.getId(), cliente);
         }
@@ -94,7 +94,7 @@ public class ClienteRepository {
         }
     }
     
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         clientes.remove(id);
     }
